+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### New features
 
 * Add a new collection: `kala.collection.mutable.MutableLinkedHashSet`.
+* Add `kala.collection.CollectionLike::toArraySeq()`
 
 ### Breaking Changes
 
@@ -19,7 +20,9 @@ Other changes:
 
 * The parameter type of `kala.collection.immutable.ImmutableArraySeq.Unsafe.wrap(Object[])` is changed to `E[]`.
 * The return type of `kala.collection.CollectionLike::toSeq()` changes to `ImmutableSeq<E>`.
+* The return type of `kala.collection.CollectionLike::toSet()` changes to `ImmutableSet<E>`.
 * `kala.collection.CollectionLike::toImmutableSeq()` is deprecated for removal and is replaced by `toSeq()`.
+* `kala.collection.CollectionLike::toImmutableSet()` is deprecated for removal and is replaced by `toSet()`.
 
 ## 0.81.0